Add linear regression to chiller validation
